More work on the functional test cases to load games
The simple functional test with 3 players (now all AI, to avoid bringing up the GUIs which takes looooooong) now cleanly terminates the game and test case is ok.
The test case with the 6 players (which currently fails due to the recruit and the legion order) would fail, but sometimes takes long (or gets even stuck??) so I have commented the actual call to "Start.main()" out for now.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4239 c5679a2a-0e0e-0410-bba2-8642b794b8c1