We now explicitly set the variant after it was loaded during the game load.
This seems to fix the issue with the game load test breaking if run together with other test cases. It's not an elegant fix, but it is probably a lot better than relying on the variant being null and in consequence being retrieved from the external source later on (which probably ends up using some static members).

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4290 c5679a2a-0e0e-0410-bba2-8642b794b8c1