Changed the initial split signature from List<String> to List<CreatureType>.
In some ways it could be List<Creature>, but that would most likely involve some larger surgery. The current change just moves the creature type lookup from the clients of IVariantHint into the implementations. This currently means lots of separate calls to the lookup methods instead of using a loop, but that can be improved further at a later stage. For now I did the change with the lowest impact (hopefully) and thanks to Eclipse's combination of problem view, quick fix and regex replacements it was pretty trivial to do.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4302 c5679a2a-0e0e-0410-bba2-8642b794b8c1