Changed default variant to use members for creature types.
All creature types used in the hinting process are now resolved at construction time, not during the usage. In theory this should be done for all variants, but for now it is only done for default since it is a bit tedious to change. Assuming that Default is used to create new variants it should show the better way of doing things, though.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4326 c5679a2a-0e0e-0410-bba2-8642b794b8c1