Moved some more methods up to the Legion base class.
This time it is addCreature(..) and removeCreature(..). The server side has a few options on that one, but I just forwarded the methods with some seemingly useful versions. Conceptually it certainly makes sense to have the methods on the game model.

I also removed unnecessary casts around those methods and the setRecruit(..) method.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4341 c5679a2a-0e0e-0410-bba2-8642b794b8c1