Removed the whole getId/fromId code from EntrySide.
It basically was just more complicated versions of the ordinal()/values() methods Java enums offer for free.

I also dropped the constraint that fromLabel(..) had to result in an attacker side, since the code would just have thrown an exception otherwise. That would have happened by now if there would be a path to get it and conceptually there doesn't seem to be a reason for the restriction.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4364 c5679a2a-0e0e-0410-bba2-8642b794b8c1