Moved a lot of battleland setup code from HexMap to MasterBoardTerrain.
Instead of having the two static Map<MasterBoardTerrain, X> in HexMap, the hex layout is now stored as normal members (i.e. plain "private final X"s) in MasterBoardTerrain. This removed a lot of static code from HexMap, and quite a few classes do not have any dependency on HexMap anymore.

The code moved was also changed from using GuiBattleHexes to the model BattleHexes. Surprisingly that didn't seem to hurt at all -- straight after managing to make the code compile I was able to run the unit tests successfully and also play a full game. Since that is not to be expected from a change this large I strongly recommend to embrace for impact ;-)

But the new code is certainly much better and I can't think of anything particular that I ought to test right now, thus this commit. A full game against 5 AIs seems a decent enough sanity test.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4365 c5679a2a-0e0e-0410-bba2-8642b794b8c1