Renamed "event" objects to action-based names.
While I used the "Action" suffix for all abstract (or supposedly abstract) classes and interfaces, I decided to go for a less verbose naming scheme on the concrete classes, e.g. having a "Summoning" as an "AddCreatureAction". I was contemplating the "SummonAction" instead, but the shorter form seemed more natural.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4416 c5679a2a-0e0e-0410-bba2-8642b794b8c1