Added mechanism to include version number in About dialog.
If a property "release.version" is set in the toplevel "build.properties" file (or the "local_build.properties"), then the value of that property will be displayed as version in the About dialog. If it is not set, "SNAPSHOT" will be used.

This is untested since I don't have a SVN working copy here, which is required to generate the build.properties file in META-INF. And I still need to fix that name clash between those two files, too.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4479 c5679a2a-0e0e-0410-bba2-8642b794b8c1