Fixed: 2905534 NPE in GameServerSide.engage
There was used getEngagement().toString() in an else case where it might (in theory) reach also if there is no engagement ongoing. In practive it is somehow questionable how this could happen, because it means someone had clicked on a hex to engage but there were no opposing legions. Perhaps an AI caching from begin of Engagements phase all engagements and now one was not existing any more because in an earlier engagement a Titan was killed? Changed message to warning so we will see if this happens again.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4622 c5679a2a-0e0e-0410-bba2-8642b794b8c1