Improvements to the handling of "when a player closes the BattleBoard window":
For a not involved player, just hide the window;
If involved, now do concede, instead of "ending the game / withdraw";
If involved and own legion has Titan, additional warning that this means one would withdraw and that this will score same way as if he would win the battle "the normal way";
If it is not own battle turn, additional information that active player has to finish his turn anyway.
If the battle active player would then just press Done to finish it, still show a warning, if Titan is offboard would still count as also dead.
(still to be done: if inactive player concedes, inform the active player that the other one has conceded, not much use to spent thought on "how to do best" - main thing: put all creatures onboard, if necessary (or leave crap offboard to get rid of it ;-))



git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4837 c5679a2a-0e0e-0410-bba2-8642b794b8c1