Giving up writing to clients after first fail, instead register to selector to get notified when it becomes writable again. Additionally sends some "tellWhatsHappening" message to the other clients.
This eliminates the following problem: When writing to one client failed, it did try and retry in 1 sec interval again and again until it recovered, or after 15 mins or so hit some network timeout. As result, it did not listen to any other client and was thus 15 mins unresponsive, thus whole game stuck.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4867 c5679a2a-0e0e-0410-bba2-8642b794b8c1