Fixed: 2884952 GetPlayers: disabled player names
For check for duplicate names and using for starting game, only consider the first N players, where N is number of players of the selected variant. Separately, store names and types of players (N+1 ... end) back to options object, so that they are preserved for next game (if one would choose another (more-player-) variant then again).


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4924 c5679a2a-0e0e-0410-bba2-8642b794b8c1