Better error handling for game starting:
Added a "systemMessage()" method. Old clients it just send a message as SYSTEM in chat, new clients it could do something else (still to be done, just added in interfaces that this method exists);
Informing via systemMessage for each client that connected;
Modified the game start failed message to tell which clients connected and which not;
Loops for informingAllEnrolled - Players changed "player null" etc. from info to warning.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4952 c5679a2a-0e0e-0410-bba2-8642b794b8c1