WebClient Socket thread (client side): for failed/rejected login requests make sure we get rid of the initial thread.
If the initial connecting thread is rejected with a failure reason (wrong password, other session still up and we need to force, etc.), we need to make sure we end that own (first) socket thread properly, even if we do not receive server's "kick out out" quick enough (that's how it worked so far, it seems). If we rely fully on server, that initial thread may get back to it's readline loop and never quit, thus making client side JVM never exiting "cleanly" by itself, i.e. depending on the TimedJVMQuit timeout.



git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4964 c5679a2a-0e0e-0410-bba2-8642b794b8c1