Fixed: 3160863 "illegal call to engage() message due to slow server response"
Client keeps now track of whether an engaging (waiting for server response) or actual engagement is already ongoing, and if so, shows a popup message instead of sending another engage. (Until now, the time window between sending engage() and receive the initial server response to it was not protected at all, and the case that an engagement is active was checked only in actOnHex, but not in actOnLegion case).

Log message on server side also tells the hex on which player attempts to engage.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4979 c5679a2a-0e0e-0410-bba2-8642b794b8c1