Fixed: 3160873 "Concede hangs game" and 3133960 "Conceding battle leads to hung game".
This is kind of a hacky solution: To get rid of the risk of hanging the game, now the concded procedure is not fully compliant with the rules (inactive player is normnally not supposed to concede). However, I consider hung games a worse thing than compliance with a minor aspect of rules, which most players are probably not even aware of, nor would they care. 

See Feature Request 3182336 "Concede of non-phasing player" for details.

Additionally this resolves the issue, that sometimes 2 times concede was needed (first concede only advanced to move phase, another needed to really finish it).


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4988 c5679a2a-0e0e-0410-bba2-8642b794b8c1