Fixed: 3184700 Wrong phase when selecting creature to move
The problem occured because the strike+strikeback phase of defenders first move phase in practice do really happen, they just are over immediately because client's doAutoStrikes does doneWithStrikes because there are no striked possible. If the attacker clicks his creatures repeatedly (prematurely), in that short timewindow when it's "his" turn the "isMyBattlePhase()" was true, and this caused a "actOnMisclick()".
Which called leaveCarryMove, which in turn sends a DoneWithStrikePhase() to server, which is the reason for the wrong phase.
This correction now skips the leaveCarryMode() if it's is not really in carry mode.

The "can't move the one creature later on" appeared on server side as a "null move": moved from e.g. X1 to X1.
I have not found any explanation how this could happen. So that part could in theory still happen.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4999 c5679a2a-0e0e-0410-bba2-8642b794b8c1