Further cleanup of usage of synchronized:
The syncdEventList is now not a Collections.synchronizedList() any more because all access is inside synchronized blocks, and most accesses are non-atomic blocks anyway.
Both lists now LinkedList instead of ArrayList because they are almost always traversed with iterators anyway.
Removed synchronized from addToEventList() because it's synchronizing on the list access anyway.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@5010 c5679a2a-0e0e-0410-bba2-8642b794b8c1