Changed the PickMarker dialog from static method, perhaps in a "loop until marker chosen", to a simple dialog with final variables.
The dialog, upon completion, initiates then the needed activity in the client: for split legion, bring up the split, or forget about it; for initial marker, send it to server, or bring up marker dialog again until one marker was chosen. So there is no "inside a loop bring up dialog, wait for it's return, and if null do it again".
This improvment means that the SocketClientThread is not hung in the wait() (and thus not responsive to the server) until the selection would be done, as it used to be.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@5013 c5679a2a-0e0e-0410-bba2-8642b794b8c1