Request-for-extra-roll feature almost ready
Server sends request for approval to clients, and client presents dialog
box for user to approve or deny, and clients send that back.
If there are no clients that can handle that, then the approval request is
send to requestor and he is told to negotiate it in chat, and then respond
to the dialog box accordingly. AI's always approve.
Still missing: server collecting the responses and/or timeout handling,
and if result is approved, handling the extra roll (perhaps can re-use the
mulligan functionality for that?)

Also the menu item is now only enabled when applicable.