Visual feedback for Done on battleboard to prevent "illegally called ..." cases
For doneWithStrikes and doneWithMoves, disable button, set waitCursor, and add
in the phase label an indicator that "Done" was done. This will be replaced
again when the server response arrives. For local games this will barely
noticable, but for remote games (public server) it might have stayed as
"can click Done" for several seconds; if user did that, caused illegally
called doneWith... errors.
Note, this is also done for the "client automatically says "Done" if there
are no strikes (especially first turn).

Also, phase label is now at the front of infoPanel;
the displayed text is different than before:
a) in "Done-but-no-reply-yet" situation it contains a ([auto-]done) text, and
b) the text is now a "verb" (moves, strikes, ..) instead of phase name.